Get input/entry working
Previously, the input wasn't fully working. This commit makes it
function as it is designed.

I eliminated one of the three types of title formats, leaving "native"
and "raw." There is some processing that happens now inside of
`render()` that was previously handled in the mapping code, but since
that processing is visual-only, it makes sense to relegate that task to
`render().`

Strings have also been promoted to full objects in the value arrays.
This is to simplify the logic when processing form saves. The challenge
is that the token field returns strings for every item in the entry
regardless if it was provided objects. Therefore, we have to do a
reverse lookup into the valid tokens table to find the keys that would
have been needed to produce the translated output. This is a lot of code
to make the token field work. It might be possible that it doesn't
provide enough of a benefit for what we are doing.

**Remaining Tasks**
 - Validate token list for each post type
 - Hook up to the API
 - Fix suggestion list (keeps items in list after they are in the value
   array)
 - Clean up UX of entering string tags (space characters,
   comma-to-insert, etc...)